refactor: extract manual balances dialog
